refactor: add user library
