+++ libs/static-dynamic-import/users/src/index.ts
@@ -1,3 +1,1 @@
-export type { User } from './lib/user.model';
-export { UserPipe } from './lib/user.pipe';
 export { default } from './lib/users.component';
