+++ libs/static-dynamic-import/users/tsconfig.lib.json
@@ -8,5 +8,9 @@
     "types": []
   },
   "exclude": ["src/**/*.spec.ts", "jest.config.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts"]
+  "include": [
+    "src/**/*.ts",
+    "../../user/src/lib/user.model.ts",
+    "../../user/src/lib/user.pipe.ts"
+  ]
 }
