+++ libs/user/README.md
@@ -1,0 +1,7 @@
+# user
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build user` to build the library.
