+++ libs/user/src/index.ts
@@ -1,0 +1,3 @@
+//export type { User } from './lib/user.model';
+export type { User } from './lib/user.model';
+export { UserPipe } from './lib/user.pipe';
