+++ tsconfig.base.json
@@ -68,6 +68,9 @@
         "libs/shared/directives/src/index.ts"
       ],
       "@angular-challenges/shared/utils": ["libs/shared/utils/src/index.ts"],
+      "@angular-challenges/static-dynamic-import/users": [
+        "libs/static-dynamic-import/users/src/index.ts"
+      ],
       "@angular-challenges/testing-table/backend": [
         "libs/testing-table/backend/src/index.ts"
       ],
@@ -77,9 +80,7 @@
       "@tomalaforge/ngrx-callstate-store": [
         "libs/shared/ngrx-callstate-store/src/index.ts"
       ],
-      "@angular-challenges/static-dynamic-import/users": [
-        "libs/static-dynamic-import/users/src/index.ts"
-      ]
+      "user": ["libs/user/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
